Create separate classes for fields with nested properties
