+++ SamplesDashboard/SamplesDashboard/Models/PullRequests.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace SamplesDashboard.Models
+{
+    public class PullRequests
+    {
+        public int totalCount;
+    }
+}
